+++ src/Server/ServerNetwork/Network/Pipeline/Outbound/ConnectionMappers/SingleEntityConnectionMapperFactory.cs
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 using System;
+using Castle.Core.Logging;
 using Sovereign.EngineUtil.Monads;
 using Sovereign.NetworkCore.Network.Infrastructure;
 using Sovereign.NetworkCore.Network.Pipeline.Outbound;
@@ -35,6 +36,8 @@ public class SingleEntityConnectionMapperFactory
         this.connectionManager = connectionManager;
     }
 
+    public ILogger Logger { private get; set; } = NullLogger.Instance;
+
     /// <summary>
     ///     Creates a connection mapper.
     /// </summary>
@@ -42,6 +45,6 @@ public class SingleEntityConnectionMapperFactory
     /// <returns>Connection mapper.</returns>
     public SingleEntityConnectionMapper Create(Func<OutboundEventInfo, Maybe<int>> mapper)
     {
-        return new SingleEntityConnectionMapper(connectionManager, mapper);
+        return new SingleEntityConnectionMapper(connectionManager, Logger, mapper);
     }
 }